requirements.txt
flask
requests
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
PyMySQL
marshmallow
flask-jwt-extended[asymmetric_crypto]
boto3